Validation

+++ src/main/java/com/example/board/board/controller/BoardController.java
@@ -2,15 +2,21 @@ package com.example.board.board.controller;
 
 import com.example.board.board.dto.*;
 import com.example.board.board.service.BoardService;
+import com.example.board.common.ResponseErrors;
 import com.example.board.common.ResponseResult;
 import com.example.board.common.ServiceResult;
 import com.example.board.member.dto.MemberLogin;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.Errors;
+import org.springframework.validation.FieldError;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.List;
 
 @RestController
@@ -18,8 +24,19 @@ import java.util.List;
 public class BoardController {
     private final BoardService boardService;
 
+    @GetMapping("/api/post/")
+    public ResponseEntity<?> getAllPost(@RequestParam int st, @RequestParam int len) {
+        ServiceResult result = boardService.getAllPost(st, len);
+        return ResponseResult.result(result);
+    }
+
     @PostMapping("/api/post/do-posting")
-    public ResponseEntity<?> doPosting(@RequestBody DoPostingModel doPostingModel, List<MultipartFile> images) throws IOException {
+    public ResponseEntity<?> doPosting(@Valid @RequestParam(name = "data") DoPostingModel doPostingModel, Errors errors, @RequestParam(required = false) List<MultipartFile> images) throws IOException {
+        List<ResponseErrors> responseErrors = new ArrayList<>();
+        if (errors.hasErrors()) {
+            errors.getAllErrors().forEach(e -> responseErrors.add(ResponseErrors.of((FieldError) e)));
+            return new ResponseEntity<>(responseErrors, HttpStatus.BAD_REQUEST);
+        }
         ServiceResult result = boardService.doPosting(doPostingModel, images);
         return ResponseResult.result(result);
     }
@@ -31,7 +48,12 @@ public class BoardController {
     }
 
     @PostMapping("/api/post/delete/{postId}")
-    public ResponseEntity<?> delete(@PathVariable Long postId, @RequestBody MemberLogin memberLogin) {
+    public ResponseEntity<?> delete(@Valid @RequestBody MemberLogin memberLogin, Errors errors, @PathVariable Long postId) {
+        List<ResponseErrors> responseErrors = new ArrayList<>();
+        if (errors.hasErrors()) {
+            errors.getAllErrors().forEach(e -> responseErrors.add(ResponseErrors.of((FieldError) e)));
+            return new ResponseEntity<>(responseErrors, HttpStatus.BAD_REQUEST);
+        }
         ServiceResult result = boardService.delete(postId, memberLogin);
         return ResponseResult.result(result);
     }
@@ -43,25 +65,45 @@ public class BoardController {
     }
 
     @PostMapping("/api/post/write-reply/{postId}")
-    public ResponseEntity<?> writeReply(@PathVariable Long postId, @RequestHeader(name = "TOKEN") String token, @RequestBody PostReplyDto postReplyDto) {
+    public ResponseEntity<?> writeReply(@RequestBody PostReplyDto postReplyDto, Errors errors, @PathVariable Long postId, @RequestHeader(name = "TOKEN") String token) {
+        List<ResponseErrors> responseErrors = new ArrayList<>();
+        if (errors.hasErrors()) {
+            errors.getAllErrors().forEach(e -> responseErrors.add(ResponseErrors.of((FieldError) e)));
+            return new ResponseEntity<>(responseErrors, HttpStatus.BAD_REQUEST);
+        }
         ServiceResult result = boardService.writeReply(postId, token, postReplyDto);
         return ResponseResult.result(result);
     }
 
     @PostMapping("/api/post/reply-delete/{replyId}")
-    public ResponseEntity<?> deleteReply(@PathVariable Long replyId, @RequestHeader(name = "TOKEN") String token, @RequestBody DeleteReplyDto deleteReplyDto) {
+    public ResponseEntity<?> deleteReply(@RequestBody DeleteReplyDto deleteReplyDto, Errors errors, @PathVariable Long replyId, @RequestHeader(name = "TOKEN") String token) {
+        List<ResponseErrors> responseErrors = new ArrayList<>();
+        if (errors.hasErrors()) {
+            errors.getAllErrors().forEach(e -> responseErrors.add(ResponseErrors.of((FieldError) e)));
+            return new ResponseEntity<>(responseErrors, HttpStatus.BAD_REQUEST);
+        }
         ServiceResult result = boardService.deleteReply(replyId, token, deleteReplyDto);
         return ResponseResult.result(result);
     }
 
     @PostMapping("/api/post/write-answer/{replyId}")
-    public ResponseEntity<?> writeAnswer(@PathVariable Long replyId, @RequestHeader(name = "TOKEN") String token, @RequestBody WriteAnswerDto writeAnswerDto) {
+    public ResponseEntity<?> writeAnswer(@RequestBody WriteAnswerDto writeAnswerDto, Errors errors, @PathVariable Long replyId, @RequestHeader(name = "TOKEN") String token) {
+        List<ResponseErrors> responseErrors = new ArrayList<>();
+        if (errors.hasErrors()) {
+            errors.getAllErrors().forEach(e -> responseErrors.add(ResponseErrors.of((FieldError) e)));
+            return new ResponseEntity<>(responseErrors, HttpStatus.BAD_REQUEST);
+        }
         ServiceResult result = boardService.writeAnswer(replyId, token, writeAnswerDto);
         return ResponseResult.result(result);
     }
 
     @PostMapping("/api/post/delete-answer/{answerId}")
-    public ResponseEntity<?> deleteAnswer(@PathVariable Long answerId, @RequestHeader(name = "TOKEN") String token, @RequestBody DeleteAnswerDto deleteAnswerDto) {
+    public ResponseEntity<?> deleteAnswer(@RequestBody DeleteAnswerDto deleteAnswerDto, Errors errors, @PathVariable Long answerId, @RequestHeader(name = "TOKEN") String token) {
+        List<ResponseErrors> responseErrors = new ArrayList<>();
+        if (errors.hasErrors()) {
+            errors.getAllErrors().forEach(e -> responseErrors.add(ResponseErrors.of((FieldError) e)));
+            return new ResponseEntity<>(responseErrors, HttpStatus.BAD_REQUEST);
+        }
         ServiceResult result = boardService.deleteAnswer(answerId, token, deleteAnswerDto);
         return ResponseResult.result(result);
     }
